Alluvial Diagrams

Alluvial Diagrams

alluvial(
  data,
  vars,
  condensationvar,
  excl = TRUE,
  marg = FALSE,
  fill = "first_variable",
  bin = "default",
  orient = "vert",
  usetitle = FALSE,
  mytitle = "Alluvial Plot"
)

Arguments

data

The data as a data frame.

vars

a string naming the variables from data that contains the values used for the Alluvial Diagram.

condensationvar

The primary variable to be used for condensation.

excl

.

marg

.

fill

A list for the argument fill for selecting the variable to be represented by color. Default is 'first_variable'.

bin

labels for the bins from low to high

orient

.

usetitle

.

mytitle

.

Value

A results object containing:

results$todo					a html

results$plot					an image

results$plot2					an image