Alluvial Plot

This tool will help you form Alluvial Diagrams (Alluvial Plots).

Value

Alluvial Plot

Super classes

jmvcore::Analysis -> ClinicoPathDescriptives::alluvialBase -> alluvialClass

Methods

Public methods

alluvialClass$clone()

Inherited methods

jmvcore::Analysis$.createImage()

jmvcore::Analysis$.createImages()

jmvcore::Analysis$.createPlotObject()

jmvcore::Analysis$.load()

jmvcore::Analysis$.render()

jmvcore::Analysis$.save()

jmvcore::Analysis$.savePart()

jmvcore::Analysis$.setCheckpoint()

jmvcore::Analysis$.setParent()

jmvcore::Analysis$.setReadDatasetHeaderSource()

jmvcore::Analysis$.setReadDatasetSource()

jmvcore::Analysis$.setResourcesPathSource()

jmvcore::Analysis$.setStatePathSource()

jmvcore::Analysis$addAddon()

jmvcore::Analysis$asProtoBuf()

jmvcore::Analysis$asSource()

jmvcore::Analysis$check()

jmvcore::Analysis$init()

jmvcore::Analysis$optionsChangedHandler()

jmvcore::Analysis$postInit()

jmvcore::Analysis$print()

jmvcore::Analysis$readDataset()

jmvcore::Analysis$run()

jmvcore::Analysis$serialize()

jmvcore::Analysis$setError()

jmvcore::Analysis$setStatus()

jmvcore::Analysis$translate()

ClinicoPathDescriptives::alluvialBase$initialize()

Method clone()

The objects of this class are cloneable with this method.

Usage

alluvialClass$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.